Move toggle meta editor to own component
